backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.2.2
drf-yasg==1.21.7
oauthlib==3.2.2
psycopg2-binary==2.9.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
django-cors-headers==4.3.1
